config tests
